provided SamplesSummary with full interface required to be used in CovarianceInterpolator
